Move unit tests on data to ast package
